% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\docType{package}
\name{admin_googleAuthR}
\alias{admin_googleAuthR}
\alias{admin_googleAuthR-package}
\title{Admin Directory API
The Admin SDK Directory API lets you view and manage enterprise resources such as users and groups, administrative notifications, security features, and more.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:32:24
filename: /Users/mark/dev/R/autoGoogleAPI/googleadmindirectoryv1.auto/R/admin_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.customer
\item https://www.googleapis.com/auth/admin.directory.customer.readonly
\item https://www.googleapis.com/auth/admin.directory.device.chromeos
\item https://www.googleapis.com/auth/admin.directory.device.chromeos.readonly
\item https://www.googleapis.com/auth/admin.directory.device.mobile
\item https://www.googleapis.com/auth/admin.directory.device.mobile.action
\item https://www.googleapis.com/auth/admin.directory.device.mobile.readonly
\item https://www.googleapis.com/auth/admin.directory.domain
\item https://www.googleapis.com/auth/admin.directory.domain.readonly
\item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.member
\item https://www.googleapis.com/auth/admin.directory.group.member.readonly
\item https://www.googleapis.com/auth/admin.directory.group.readonly
\item https://www.googleapis.com/auth/admin.directory.notifications
\item https://www.googleapis.com/auth/admin.directory.orgunit
\item https://www.googleapis.com/auth/admin.directory.orgunit.readonly
\item https://www.googleapis.com/auth/admin.directory.resource.calendar
\item https://www.googleapis.com/auth/admin.directory.resource.calendar.readonly
\item https://www.googleapis.com/auth/admin.directory.rolemanagement
\item https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly
\item https://www.googleapis.com/auth/admin.directory.user
\item https://www.googleapis.com/auth/admin.directory.user.alias
\item https://www.googleapis.com/auth/admin.directory.user.alias.readonly
\item https://www.googleapis.com/auth/admin.directory.user.readonly
\item https://www.googleapis.com/auth/admin.directory.user.security
\item https://www.googleapis.com/auth/admin.directory.userschema
\item https://www.googleapis.com/auth/admin.directory.userschema.readonly
}
}

