% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Alias}
\alias{Alias}
\title{Admin Directory API Objects 
The Admin SDK Directory API lets you view and manage enterprise resources such as users and groups, administrative notifications, security features, and more.}
\usage{
Alias(alias = NULL, etag = NULL, id = NULL, primaryEmail = NULL)
}
\arguments{
\item{alias}{A alias email}

\item{etag}{ETag of the resource}

\item{id}{Unique id of the group (Read-only) Unique id of the user (Read-only)}

\item{primaryEmail}{Group's primary email (Read-only) User's primary email (Read-only)}
}
\value{
Alias object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:32:26
filename: /Users/mark/dev/R/autoGoogleAPI/googleadmindirectoryv1.auto/R/admin_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Alias Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Alias object in Directory API.
}

