% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Aliases}
\alias{Aliases}
\title{Aliases Object}
\usage{
Aliases(aliases = NULL, etag = NULL)
}
\arguments{
\item{aliases}{List of alias objects}

\item{etag}{ETag of the resource}
}
\value{
Aliases object
}
\description{
Aliases Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template to list aliases in Directory API.
}
