% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Asp}
\alias{Asp}
\title{Asp Object}
\usage{
Asp(codeId = NULL, creationTime = NULL, etag = NULL,
  lastTimeUsed = NULL, name = NULL, userKey = NULL)
}
\arguments{
\item{codeId}{The unique ID of the ASP}

\item{creationTime}{The time when the ASP was created}

\item{etag}{ETag of the ASP}

\item{lastTimeUsed}{The time when the ASP was last used}

\item{name}{The name of the application that the user, represented by their userId, entered when the ASP was created}

\item{userKey}{The unique ID of the user who issued the ASP}
}
\value{
Asp object
}
\description{
Asp Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The template that returns individual ASP (Access Code) data.
}
