% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Asps}
\alias{Asps}
\title{Asps Object}
\usage{
Asps(etag = NULL, items = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{A list of ASP resources}
}
\value{
Asps object
}
\description{
Asps Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

