% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{CalendarResource}
\alias{CalendarResource}
\title{CalendarResource Object}
\usage{
CalendarResource(etags = NULL, resourceDescription = NULL,
  resourceEmail = NULL, resourceId = NULL, resourceName = NULL,
  resourceType = NULL)
}
\arguments{
\item{etags}{ETag of the resource}

\item{resourceDescription}{The brief description of the calendar resource}

\item{resourceEmail}{The read-only email ID for the calendar resource}

\item{resourceId}{The unique ID for the calendar resource}

\item{resourceName}{The name of the calendar resource}

\item{resourceType}{The type of the calendar resource}
}
\value{
CalendarResource object
}
\description{
CalendarResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Calendar Resource object in Directory API.
}
\seealso{
Other CalendarResource functions: \code{\link{directory.resources.calendars.insert}},
  \code{\link{directory.resources.calendars.patch}},
  \code{\link{directory.resources.calendars.update}}
}
