% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{CalendarResources}
\alias{CalendarResources}
\title{CalendarResources Object}
\usage{
CalendarResources(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{The CalendarResources in this page of results}

\item{nextPageToken}{The continuation token, used to page through large result sets}
}
\value{
CalendarResources object
}
\description{
CalendarResources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Calendar Resource List Response object in Directory API.
}

