% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{channels.stop}
\alias{channels.stop}
\title{Stop watching resources through this channel}
\usage{
channels.stop(Channel)
}
\arguments{
\item{Channel}{The \link{Channel} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.user
\item https://www.googleapis.com/auth/admin.directory.user.alias
\item https://www.googleapis.com/auth/admin.directory.user.alias.readonly
\item https://www.googleapis.com/auth/admin.directory.user.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.user, https://www.googleapis.com/auth/admin.directory.user.alias, https://www.googleapis.com/auth/admin.directory.user.alias.readonly, https://www.googleapis.com/auth/admin.directory.user.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other Channel functions: \code{\link{Channel.params}},
  \code{\link{Channel}},
  \code{\link{directory.users.watch}}
}
