% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{ChromeOsDevice.activeTimeRanges}
\alias{ChromeOsDevice.activeTimeRanges}
\title{ChromeOsDevice.activeTimeRanges Object}
\usage{
ChromeOsDevice.activeTimeRanges()
}
\value{
ChromeOsDevice.activeTimeRanges object
}
\description{
ChromeOsDevice.activeTimeRanges Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ChromeOsDevice functions: \code{\link{ChromeOsDevice.recentUsers}},
  \code{\link{ChromeOsDevice}},
  \code{\link{directory.chromeosdevices.patch}},
  \code{\link{directory.chromeosdevices.update}}
}
