% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{ChromeOsDevice.recentUsers}
\alias{ChromeOsDevice.recentUsers}
\title{ChromeOsDevice.recentUsers Object}
\usage{
ChromeOsDevice.recentUsers()
}
\value{
ChromeOsDevice.recentUsers object
}
\description{
ChromeOsDevice.recentUsers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ChromeOsDevice functions: \code{\link{ChromeOsDevice.activeTimeRanges}},
  \code{\link{ChromeOsDevice}},
  \code{\link{directory.chromeosdevices.patch}},
  \code{\link{directory.chromeosdevices.update}}
}
