% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{ChromeOsDeviceAction}
\alias{ChromeOsDeviceAction}
\title{ChromeOsDeviceAction Object}
\usage{
ChromeOsDeviceAction(action = NULL, deprovisionReason = NULL)
}
\arguments{
\item{action}{Action to be taken on the ChromeOs Device}

\item{deprovisionReason}{No description}
}
\value{
ChromeOsDeviceAction object
}
\description{
ChromeOsDeviceAction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON request template for firing actions on ChromeOs Device in Directory Devices API.
}
\seealso{
Other ChromeOsDeviceAction functions: \code{\link{directory.chromeosdevices.action}}
}
