% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{ChromeOsDevices}
\alias{ChromeOsDevices}
\title{ChromeOsDevices Object}
\usage{
ChromeOsDevices(chromeosdevices = NULL, etag = NULL, nextPageToken = NULL)
}
\arguments{
\item{chromeosdevices}{List of Chrome OS Device objects}

\item{etag}{ETag of the resource}

\item{nextPageToken}{Token used to access next page of this result}
}
\value{
ChromeOsDevices object
}
\description{
ChromeOsDevices Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List Chrome OS Devices operation in Directory API.
}
