% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Customer}
\alias{Customer}
\title{Customer Object}
\usage{
Customer(alternateEmail = NULL, customerCreationTime = NULL,
  customerDomain = NULL, etag = NULL, id = NULL, language = NULL,
  phoneNumber = NULL, postalAddress = NULL)
}
\arguments{
\item{alternateEmail}{The customer's secondary contact email address}

\item{customerCreationTime}{The customer's creation time (Readonly)}

\item{customerDomain}{The customer's primary domain name string}

\item{etag}{ETag of the resource}

\item{id}{The unique ID for the customer's Google account}

\item{language}{The customer's ISO 639-2 language code}

\item{phoneNumber}{The customer's contact phone number in E}

\item{postalAddress}{The customer's postal address information}
}
\value{
Customer object
}
\description{
Customer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Customer Resource object in Directory API.
}
\seealso{
Other Customer functions: \code{\link{directory.customers.patch}},
  \code{\link{directory.customers.update}}
}

