% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{CustomerPostalAddress}
\alias{CustomerPostalAddress}
\title{CustomerPostalAddress Object}
\usage{
CustomerPostalAddress(addressLine1 = NULL, addressLine2 = NULL,
  addressLine3 = NULL, contactName = NULL, countryCode = NULL,
  locality = NULL, organizationName = NULL, postalCode = NULL,
  region = NULL)
}
\arguments{
\item{addressLine1}{A customer's physical address}

\item{addressLine2}{Address line 2 of the address}

\item{addressLine3}{Address line 3 of the address}

\item{contactName}{The customer contact's name}

\item{countryCode}{This is a required property}

\item{locality}{Name of the locality}

\item{organizationName}{The company or company division name}

\item{postalCode}{The postal code}

\item{region}{Name of the region}
}
\value{
CustomerPostalAddress object
}
\description{
CustomerPostalAddress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for postal address of a customer.
}
