% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.chromeosdevices.action}
\alias{directory.chromeosdevices.action}
\title{Take action on Chrome OS Device}
\usage{
directory.chromeosdevices.action(ChromeOsDeviceAction, customerId, resourceId)
}
\arguments{
\item{ChromeOsDeviceAction}{The \link{ChromeOsDeviceAction} object to pass to this method}

\item{customerId}{Immutable id of the Google Apps account}

\item{resourceId}{Immutable id of Chrome OS Device}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.device.chromeos
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.device.chromeos)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other ChromeOsDeviceAction functions: \code{\link{ChromeOsDeviceAction}}
}
