% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.chromeosdevices.list}
\alias{directory.chromeosdevices.list}
\title{Retrieve all Chrome OS Devices of a customer (paginated)}
\usage{
directory.chromeosdevices.list(customerId, maxResults = NULL,
  orderBy = NULL, pageToken = NULL, projection = NULL, query = NULL,
  sortOrder = NULL)
}
\arguments{
\item{customerId}{Immutable id of the Google Apps account}

\item{maxResults}{Maximum number of results to return}

\item{orderBy}{Column to use for sorting results}

\item{pageToken}{Token to specify next page in the list}

\item{projection}{Restrict information returned to a set of selected fields}

\item{query}{Search string in the format given at http://support}

\item{sortOrder}{Whether to return results in ascending or descending order}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.device.chromeos
\item https://www.googleapis.com/auth/admin.directory.device.chromeos.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.device.chromeos, https://www.googleapis.com/auth/admin.directory.device.chromeos.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

