% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.chromeosdevices.patch}
\alias{directory.chromeosdevices.patch}
\title{Update Chrome OS Device. This method supports patch semantics.}
\usage{
directory.chromeosdevices.patch(ChromeOsDevice, customerId, deviceId,
  projection = NULL)
}
\arguments{
\item{ChromeOsDevice}{The \link{ChromeOsDevice} object to pass to this method}

\item{customerId}{Immutable id of the Google Apps account}

\item{deviceId}{Immutable id of Chrome OS Device}

\item{projection}{Restrict information returned to a set of selected fields}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.device.chromeos
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.device.chromeos)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other ChromeOsDevice functions: \code{\link{ChromeOsDevice.activeTimeRanges}},
  \code{\link{ChromeOsDevice.recentUsers}},
  \code{\link{ChromeOsDevice}},
  \code{\link{directory.chromeosdevices.update}}
}

