% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.customers.get}
\alias{directory.customers.get}
\title{Retrieves a customer.}
\usage{
directory.customers.get(customerKey)
}
\arguments{
\item{customerKey}{Id of the customer to be retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.customer
\item https://www.googleapis.com/auth/admin.directory.customer.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.customer, https://www.googleapis.com/auth/admin.directory.customer.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

