% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.customers.update}
\alias{directory.customers.update}
\title{Updates a customer.}
\usage{
directory.customers.update(Customer, customerKey)
}
\arguments{
\item{Customer}{The \link{Customer} object to pass to this method}

\item{customerKey}{Id of the customer to be updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.customer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.customer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other Customer functions: \code{\link{Customer}},
  \code{\link{directory.customers.patch}}
}

