% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.domainAliases.get}
\alias{directory.domainAliases.get}
\title{Retrieves a domain alias of the customer.}
\usage{
directory.domainAliases.get(customer, domainAliasName)
}
\arguments{
\item{customer}{Immutable id of the Google Apps account}

\item{domainAliasName}{Name of domain alias to be retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.domain
\item https://www.googleapis.com/auth/admin.directory.domain.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.domain, https://www.googleapis.com/auth/admin.directory.domain.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}
