% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.domainAliases.insert}
\alias{directory.domainAliases.insert}
\title{Inserts a Domain alias of the customer.}
\usage{
directory.domainAliases.insert(DomainAlias, customer)
}
\arguments{
\item{DomainAlias}{The \link{DomainAlias} object to pass to this method}

\item{customer}{Immutable id of the Google Apps account}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.domain
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.domain)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other DomainAlias functions: \code{\link{DomainAlias}}
}
