% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.domainAliases.list}
\alias{directory.domainAliases.list}
\title{Lists the domain aliases of the customer.}
\usage{
directory.domainAliases.list(customer, parentDomainName = NULL)
}
\arguments{
\item{customer}{Immutable id of the Google Apps account}

\item{parentDomainName}{Name of the parent domain for which domain aliases are to be fetched}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.domain
\item https://www.googleapis.com/auth/admin.directory.domain.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.domain, https://www.googleapis.com/auth/admin.directory.domain.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}
