% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.groups.list}
\alias{directory.groups.list}
\title{Retrieve all groups in a domain (paginated)}
\usage{
directory.groups.list(customer = NULL, domain = NULL, maxResults = NULL,
  pageToken = NULL, userKey = NULL)
}
\arguments{
\item{customer}{Immutable id of the Google Apps account}

\item{domain}{Name of the domain}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Token to specify next page in the list}

\item{userKey}{Email or immutable Id of the user if only those groups are to be listed, the given user is a member of}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.group, https://www.googleapis.com/auth/admin.directory.group.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

