% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.members.list}
\alias{directory.members.list}
\title{Retrieve all members in a group (paginated)}
\usage{
directory.members.list(groupKey, maxResults = NULL, pageToken = NULL,
  roles = NULL)
}
\arguments{
\item{groupKey}{Email or immutable Id of the group}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Token to specify next page in the list}

\item{roles}{Comma separated role values to filter list results on}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.member
\item https://www.googleapis.com/auth/admin.directory.group.member.readonly
\item https://www.googleapis.com/auth/admin.directory.group.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.group, https://www.googleapis.com/auth/admin.directory.group.member, https://www.googleapis.com/auth/admin.directory.group.member.readonly, https://www.googleapis.com/auth/admin.directory.group.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

