% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.members.update}
\alias{directory.members.update}
\title{Update membership of a user in the specified group.}
\usage{
directory.members.update(Member, groupKey, memberKey)
}
\arguments{
\item{Member}{The \link{Member} object to pass to this method}

\item{groupKey}{Email or immutable Id of the group}

\item{memberKey}{Email or immutable Id of the user}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.member
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.group, https://www.googleapis.com/auth/admin.directory.group.member)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other Member functions: \code{\link{Member}},
  \code{\link{directory.members.insert}},
  \code{\link{directory.members.patch}}
}
