% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.mobiledevices.action}
\alias{directory.mobiledevices.action}
\title{Take action on Mobile Device}
\usage{
directory.mobiledevices.action(MobileDeviceAction, customerId, resourceId)
}
\arguments{
\item{MobileDeviceAction}{The \link{MobileDeviceAction} object to pass to this method}

\item{customerId}{Immutable id of the Google Apps account}

\item{resourceId}{Immutable id of Mobile Device}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.device.mobile
\item https://www.googleapis.com/auth/admin.directory.device.mobile.action
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.device.mobile, https://www.googleapis.com/auth/admin.directory.device.mobile.action)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other MobileDeviceAction functions: \code{\link{MobileDeviceAction}}
}

