% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.mobiledevices.delete}
\alias{directory.mobiledevices.delete}
\title{Delete Mobile Device}
\usage{
directory.mobiledevices.delete(customerId, resourceId)
}
\arguments{
\item{customerId}{Immutable id of the Google Apps account}

\item{resourceId}{Immutable id of Mobile Device}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.device.mobile
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.device.mobile)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}
