% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.mobiledevices.get}
\alias{directory.mobiledevices.get}
\title{Retrieve Mobile Device}
\usage{
directory.mobiledevices.get(customerId, resourceId, projection = NULL)
}
\arguments{
\item{customerId}{Immutable id of the Google Apps account}

\item{resourceId}{Immutable id of Mobile Device}

\item{projection}{Restrict information returned to a set of selected fields}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.device.mobile
\item https://www.googleapis.com/auth/admin.directory.device.mobile.action
\item https://www.googleapis.com/auth/admin.directory.device.mobile.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.device.mobile, https://www.googleapis.com/auth/admin.directory.device.mobile.action, https://www.googleapis.com/auth/admin.directory.device.mobile.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

