% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.notifications.list}
\alias{directory.notifications.list}
\title{Retrieves a list of notifications.}
\usage{
directory.notifications.list(customer, language = NULL, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{customer}{The unique ID for the customer's Google account}

\item{language}{The ISO 639-1 code of the language notifications are returned in}

\item{maxResults}{Maximum number of notifications to return per page}

\item{pageToken}{The token to specify the page of results to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.notifications
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.notifications)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

