% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.notifications.patch}
\alias{directory.notifications.patch}
\title{Updates a notification. This method supports patch semantics.}
\usage{
directory.notifications.patch(Notification, customer, notificationId)
}
\arguments{
\item{Notification}{The \link{Notification} object to pass to this method}

\item{customer}{The unique ID for the customer's Google account}

\item{notificationId}{The unique ID of the notification}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.notifications
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.notifications)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other Notification functions: \code{\link{Notification}},
  \code{\link{directory.notifications.update}}
}

