% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.orgunits.list}
\alias{directory.orgunits.list}
\title{Retrieve all Organization Units}
\usage{
directory.orgunits.list(customerId, orgUnitPath = NULL, type = NULL)
}
\arguments{
\item{customerId}{Immutable id of the Google Apps account}

\item{orgUnitPath}{the URL-encoded organization unit's path or its Id}

\item{type}{Whether to return all sub-organizations or just immediate children}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.orgunit
\item https://www.googleapis.com/auth/admin.directory.orgunit.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.orgunit, https://www.googleapis.com/auth/admin.directory.orgunit.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}
