% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.resources.calendars.list}
\alias{directory.resources.calendars.list}
\title{Retrieves a list of calendar resources for an account.}
\usage{
directory.resources.calendars.list(customer, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{customer}{The unique ID for the customer's Google account}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Token to specify the next page in the list}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.resource.calendar
\item https://www.googleapis.com/auth/admin.directory.resource.calendar.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.resource.calendar, https://www.googleapis.com/auth/admin.directory.resource.calendar.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

