% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.resources.calendars.update}
\alias{directory.resources.calendars.update}
\title{Updates a calendar resource.}
\usage{
directory.resources.calendars.update(CalendarResource, customer,
  calendarResourceId)
}
\arguments{
\item{CalendarResource}{The \link{CalendarResource} object to pass to this method}

\item{customer}{The unique ID for the customer's Google account}

\item{calendarResourceId}{The unique ID of the calendar resource to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.resource.calendar
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.resource.calendar)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other CalendarResource functions: \code{\link{CalendarResource}},
  \code{\link{directory.resources.calendars.insert}},
  \code{\link{directory.resources.calendars.patch}}
}

