% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.roleAssignments.get}
\alias{directory.roleAssignments.get}
\title{Retrieve a role assignment.}
\usage{
directory.roleAssignments.get(customer, roleAssignmentId)
}
\arguments{
\item{customer}{Immutable ID of the Google Apps account}

\item{roleAssignmentId}{Immutable ID of the role assignment}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.rolemanagement
\item https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.rolemanagement, https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

