% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.roleAssignments.list}
\alias{directory.roleAssignments.list}
\title{Retrieves a paginated list of all roleAssignments.}
\usage{
directory.roleAssignments.list(customer, maxResults = NULL,
  pageToken = NULL, roleId = NULL, userKey = NULL)
}
\arguments{
\item{customer}{Immutable ID of the Google Apps account}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Token to specify the next page in the list}

\item{roleId}{Immutable ID of a role}

\item{userKey}{The user's primary email address, alias email address, or unique user ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.rolemanagement
\item https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.rolemanagement, https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}

