% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.roles.list}
\alias{directory.roles.list}
\title{Retrieves a paginated list of all the roles in a domain.}
\usage{
directory.roles.list(customer, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{customer}{Immutable id of the Google Apps account}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Token to specify the next page in the list}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.rolemanagement
\item https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.rolemanagement, https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}
}
