% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.roles.update}
\alias{directory.roles.update}
\title{Updates a role.}
\usage{
directory.roles.update(Role, customer, roleId)
}
\arguments{
\item{Role}{The \link{Role} object to pass to this method}

\item{customer}{Immutable ID of the Google Apps account}

\item{roleId}{Immutable ID of the role}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.rolemanagement
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.rolemanagement)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other Role functions: \code{\link{Role.rolePrivileges}},
  \code{\link{Role}}, \code{\link{directory.roles.insert}},
  \code{\link{directory.roles.patch}}
}
