% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.users.makeAdmin}
\alias{directory.users.makeAdmin}
\title{change admin status of a user}
\usage{
directory.users.makeAdmin(UserMakeAdmin, userKey)
}
\arguments{
\item{UserMakeAdmin}{The \link{UserMakeAdmin} object to pass to this method}

\item{userKey}{Email or immutable Id of the user as admin}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.user
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.user)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other UserMakeAdmin functions: \code{\link{UserMakeAdmin}}
}
