% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.users.undelete}
\alias{directory.users.undelete}
\title{Undelete a deleted user}
\usage{
directory.users.undelete(UserUndelete, userKey)
}
\arguments{
\item{UserUndelete}{The \link{UserUndelete} object to pass to this method}

\item{userKey}{The immutable id of the user}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.user
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.user)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other UserUndelete functions: \code{\link{UserUndelete}}
}
