% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{directory.users.watch}
\alias{directory.users.watch}
\title{Watch for changes in users list}
\usage{
directory.users.watch(Channel, customFieldMask = NULL, customer = NULL,
  domain = NULL, event = NULL, maxResults = NULL, orderBy = NULL,
  pageToken = NULL, projection = NULL, query = NULL, showDeleted = NULL,
  sortOrder = NULL, viewType = NULL)
}
\arguments{
\item{Channel}{The \link{Channel} object to pass to this method}

\item{customFieldMask}{Comma-separated list of schema names}

\item{customer}{Immutable id of the Google Apps account}

\item{domain}{Name of the domain}

\item{event}{Event on which subscription is intended (if subscribing)}

\item{maxResults}{Maximum number of results to return}

\item{orderBy}{Column to use for sorting results}

\item{pageToken}{Token to specify next page in the list}

\item{projection}{What subset of fields to fetch for this user}

\item{query}{Query string search}

\item{showDeleted}{If set to true retrieves the list of deleted users}

\item{sortOrder}{Whether to return results in ascending or descending order}

\item{viewType}{Whether to fetch the ADMIN_VIEW or DOMAIN_PUBLIC view of the user}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.directory.user
\item https://www.googleapis.com/auth/admin.directory.user.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.directory.user, https://www.googleapis.com/auth/admin.directory.user.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/directory/}{Google Documentation}

Other Channel functions: \code{\link{Channel.params}},
  \code{\link{Channel}}, \code{\link{channels.stop}}
}

