% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{DomainAlias}
\alias{DomainAlias}
\title{DomainAlias Object}
\usage{
DomainAlias(creationTime = NULL, domainAliasName = NULL, etag = NULL,
  parentDomainName = NULL, verified = NULL)
}
\arguments{
\item{creationTime}{The creation time of the domain alias}

\item{domainAliasName}{The domain alias name}

\item{etag}{ETag of the resource}

\item{parentDomainName}{The parent domain name that the domain alias is associated with}

\item{verified}{Indicates the verification state of a domain alias}
}
\value{
DomainAlias object
}
\description{
DomainAlias Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Domain Alias object in Directory API.
}
\seealso{
Other DomainAlias functions: \code{\link{directory.domainAliases.insert}}
}

