% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{DomainAliases}
\alias{DomainAliases}
\title{DomainAliases Object}
\usage{
DomainAliases(domainAliases = NULL, etag = NULL)
}
\arguments{
\item{domainAliases}{List of domain alias objects}

\item{etag}{ETag of the resource}
}
\value{
DomainAliases object
}
\description{
DomainAliases Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template to list domain aliases in Directory API.
}

