% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Domains}
\alias{Domains}
\title{Domains Object}
\usage{
Domains(creationTime = NULL, domainAliases = NULL, domainName = NULL,
  etag = NULL, isPrimary = NULL, verified = NULL)
}
\arguments{
\item{creationTime}{Creation time of the domain}

\item{domainAliases}{List of domain alias objects}

\item{domainName}{The domain name of the customer}

\item{etag}{ETag of the resource}

\item{isPrimary}{Indicates if the domain is a primary domain (Read-only)}

\item{verified}{Indicates the verification state of a domain}
}
\value{
Domains object
}
\description{
Domains Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Domain object in Directory API.
}
\seealso{
Other Domains functions: \code{\link{directory.domains.insert}}
}
