% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Domains2}
\alias{Domains2}
\title{Domains2 Object}
\usage{
Domains2(domains = NULL, etag = NULL)
}
\arguments{
\item{domains}{List of domain objects}

\item{etag}{ETag of the resource}
}
\value{
Domains2 object
}
\description{
Domains2 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template to list Domains in Directory API.
}
