% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Group}
\alias{Group}
\title{Group Object}
\usage{
Group(adminCreated = NULL, aliases = NULL, description = NULL,
  directMembersCount = NULL, email = NULL, etag = NULL, id = NULL,
  name = NULL, nonEditableAliases = NULL)
}
\arguments{
\item{adminCreated}{Is the group created by admin (Read-only) *}

\item{aliases}{List of aliases (Read-only)}

\item{description}{Description of the group}

\item{directMembersCount}{Group direct members count}

\item{email}{Email of Group}

\item{etag}{ETag of the resource}

\item{id}{Unique identifier of Group (Read-only)}

\item{name}{Group name}

\item{nonEditableAliases}{List of non editable aliases (Read-only)}
}
\value{
Group object
}
\description{
Group Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Group resource in Directory API.
}
\seealso{
Other Group functions: \code{\link{directory.groups.insert}},
  \code{\link{directory.groups.patch}},
  \code{\link{directory.groups.update}}
}
