% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Groups}
\alias{Groups}
\title{Groups Object}
\usage{
Groups(etag = NULL, groups = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{groups}{List of group objects}

\item{nextPageToken}{Token used to access next page of this result}
}
\value{
Groups object
}
\description{
Groups Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List Groups operation in Directory API.
}

