% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Member}
\alias{Member}
\title{Member Object}
\usage{
Member(email = NULL, etag = NULL, id = NULL, role = NULL,
  status = NULL, type = NULL)
}
\arguments{
\item{email}{Email of member (Read-only)}

\item{etag}{ETag of the resource}

\item{id}{Unique identifier of customer member (Read-only) Unique identifier of group (Read-only) Unique identifier of member (Read-only)}

\item{role}{Role of member}

\item{status}{Status of member (Immutable)}

\item{type}{Type of member (Immutable)}
}
\value{
Member object
}
\description{
Member Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Member resource in Directory API.
}
\seealso{
Other Member functions: \code{\link{directory.members.insert}},
  \code{\link{directory.members.patch}},
  \code{\link{directory.members.update}}
}

