% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Members}
\alias{Members}
\title{Members Object}
\usage{
Members(etag = NULL, members = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{members}{List of member objects}

\item{nextPageToken}{Token used to access next page of this result}
}
\value{
Members object
}
\description{
Members Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List Members operation in Directory API.
}

