% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{MobileDevice.applications}
\alias{MobileDevice.applications}
\title{MobileDevice.applications Object}
\usage{
MobileDevice.applications()
}
\value{
MobileDevice.applications object
}
\description{
MobileDevice.applications Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other MobileDevice functions: \code{\link{MobileDevice}}
}
