% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{MobileDevice}
\alias{MobileDevice}
\title{MobileDevice Object}
\usage{
MobileDevice(MobileDevice.applications = NULL, adbStatus = NULL,
  applications = NULL, basebandVersion = NULL, bootloaderVersion = NULL,
  brand = NULL, buildNumber = NULL, defaultLanguage = NULL,
  developerOptionsStatus = NULL, deviceCompromisedStatus = NULL,
  deviceId = NULL, devicePasswordStatus = NULL, email = NULL,
  encryptionStatus = NULL, etag = NULL, firstSync = NULL,
  hardware = NULL, hardwareId = NULL, imei = NULL, kernelVersion = NULL,
  lastSync = NULL, managedAccountIsOnOwnerProfile = NULL,
  manufacturer = NULL, meid = NULL, model = NULL, name = NULL,
  networkOperator = NULL, os = NULL, otherAccountsInfo = NULL,
  privilege = NULL, releaseVersion = NULL, resourceId = NULL,
  securityPatchLevel = NULL, serialNumber = NULL, status = NULL,
  supportsWorkProfile = NULL, type = NULL, unknownSourcesStatus = NULL,
  userAgent = NULL, wifiMacAddress = NULL)
}
\arguments{
\item{MobileDevice.applications}{The \link{MobileDevice.applications} object or list of objects}

\item{adbStatus}{Adb (USB debugging) enabled or disabled on device (Read-only)}

\item{applications}{List of applications installed on Mobile Device}

\item{basebandVersion}{Mobile Device Baseband version (Read-only)}

\item{bootloaderVersion}{Mobile Device Bootloader version (Read-only)}

\item{brand}{Mobile Device Brand (Read-only)}

\item{buildNumber}{Mobile Device Build number (Read-only)}

\item{defaultLanguage}{The default locale used on the Mobile Device (Read-only)}

\item{developerOptionsStatus}{Developer options enabled or disabled on device (Read-only)}

\item{deviceCompromisedStatus}{Mobile Device compromised status (Read-only)}

\item{deviceId}{Mobile Device serial number (Read-only)}

\item{devicePasswordStatus}{DevicePasswordStatus (Read-only)}

\item{email}{List of owner user's email addresses (Read-only)}

\item{encryptionStatus}{Mobile Device Encryption Status (Read-only)}

\item{etag}{ETag of the resource}

\item{firstSync}{Date and time the device was first synchronized with the policy settings in the Google Apps administrator control panel (Read-only)}

\item{hardware}{Mobile Device Hardware (Read-only)}

\item{hardwareId}{Mobile Device Hardware Id (Read-only)}

\item{imei}{Mobile Device IMEI number (Read-only)}

\item{kernelVersion}{Mobile Device Kernel version (Read-only)}

\item{lastSync}{Date and time the device was last synchronized with the policy settings in the Google Apps administrator control panel (Read-only)}

\item{managedAccountIsOnOwnerProfile}{Boolean indicating if this account is on owner/primary profile or not (Read-only)}

\item{manufacturer}{Mobile Device manufacturer (Read-only)}

\item{meid}{Mobile Device MEID number (Read-only)}

\item{model}{Name of the model of the device}

\item{name}{List of owner user's names (Read-only)}

\item{networkOperator}{Mobile Device mobile or network operator (if available) (Read-only)}

\item{os}{Name of the mobile operating system}

\item{otherAccountsInfo}{List of accounts added on device (Read-only)}

\item{privilege}{DMAgentPermission (Read-only)}

\item{releaseVersion}{Mobile Device release version version (Read-only)}

\item{resourceId}{Unique identifier of Mobile Device (Read-only)}

\item{securityPatchLevel}{Mobile Device Security patch level (Read-only)}

\item{serialNumber}{Mobile Device SSN or Serial Number (Read-only)}

\item{status}{Status of the device (Read-only)}

\item{supportsWorkProfile}{Work profile supported on device (Read-only)}

\item{type}{The type of device (Read-only)}

\item{unknownSourcesStatus}{Unknown sources enabled or disabled on device (Read-only)}

\item{userAgent}{Mobile Device user agent}

\item{wifiMacAddress}{Mobile Device WiFi MAC address (Read-only)}
}
\value{
MobileDevice object
}
\description{
MobileDevice Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Mobile Device resource in Directory API.
}
\seealso{
Other MobileDevice functions: \code{\link{MobileDevice.applications}}
}
