% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{MobileDeviceAction}
\alias{MobileDeviceAction}
\title{MobileDeviceAction Object}
\usage{
MobileDeviceAction(action = NULL)
}
\arguments{
\item{action}{Action to be taken on the Mobile Device}
}
\value{
MobileDeviceAction object
}
\description{
MobileDeviceAction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON request template for firing commands on Mobile Device in Directory Devices API.
}
\seealso{
Other MobileDeviceAction functions: \code{\link{directory.mobiledevices.action}}
}

