% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{MobileDevices}
\alias{MobileDevices}
\title{MobileDevices Object}
\usage{
MobileDevices(etag = NULL, mobiledevices = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{mobiledevices}{List of Mobile Device objects}

\item{nextPageToken}{Token used to access next page of this result}
}
\value{
MobileDevices object
}
\description{
MobileDevices Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List Mobile Devices operation in Directory API.
}
