% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Notification}
\alias{Notification}
\title{Notification Object}
\usage{
Notification(body = NULL, etag = NULL, fromAddress = NULL,
  isUnread = NULL, notificationId = NULL, sendTime = NULL,
  subject = NULL)
}
\arguments{
\item{body}{Body of the notification (Read-only)}

\item{etag}{ETag of the resource}

\item{fromAddress}{Address from which the notification is received (Read-only)}

\item{isUnread}{Boolean indicating whether the notification is unread or not}

\item{notificationId}{No description}

\item{sendTime}{Time at which notification was sent (Read-only)}

\item{subject}{Subject of the notification (Read-only)}
}
\value{
Notification object
}
\description{
Notification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for a notification resource.
}
\seealso{
Other Notification functions: \code{\link{directory.notifications.patch}},
  \code{\link{directory.notifications.update}}
}
