% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Notifications}
\alias{Notifications}
\title{Notifications Object}
\usage{
Notifications(etag = NULL, items = NULL, nextPageToken = NULL,
  unreadNotificationsCount = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{List of notifications in this page}

\item{nextPageToken}{Token for fetching the next page of notifications}

\item{unreadNotificationsCount}{Number of unread notification for the domain}
}
\value{
Notifications object
}
\description{
Notifications Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for notifications list response.
}
