% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{OrgUnit}
\alias{OrgUnit}
\title{OrgUnit Object}
\usage{
OrgUnit(blockInheritance = NULL, description = NULL, etag = NULL,
  name = NULL, orgUnitId = NULL, orgUnitPath = NULL,
  parentOrgUnitId = NULL, parentOrgUnitPath = NULL)
}
\arguments{
\item{blockInheritance}{Should block inheritance}

\item{description}{Description of OrgUnit}

\item{etag}{ETag of the resource}

\item{name}{Name of OrgUnit}

\item{orgUnitId}{Id of OrgUnit}

\item{orgUnitPath}{Path of OrgUnit}

\item{parentOrgUnitId}{Id of parent OrgUnit}

\item{parentOrgUnitPath}{Path of parent OrgUnit}
}
\value{
OrgUnit object
}
\description{
OrgUnit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Org Unit resource in Directory API.
}
\seealso{
Other OrgUnit functions: \code{\link{directory.orgunits.insert}},
  \code{\link{directory.orgunits.patch}},
  \code{\link{directory.orgunits.update}}
}
