% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{OrgUnits}
\alias{OrgUnits}
\title{OrgUnits Object}
\usage{
OrgUnits(etag = NULL, organizationUnits = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{organizationUnits}{List of user objects}
}
\value{
OrgUnits object
}
\description{
OrgUnits Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List Organization Units operation in Directory API.
}

