% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Privilege}
\alias{Privilege}
\title{Privilege Object}
\usage{
Privilege(childPrivileges = NULL, etag = NULL, isOuScopable = NULL,
  privilegeName = NULL, serviceId = NULL, serviceName = NULL)
}
\arguments{
\item{childPrivileges}{A list of child privileges}

\item{etag}{ETag of the resource}

\item{isOuScopable}{If the privilege can be restricted to an organization unit}

\item{privilegeName}{The name of the privilege}

\item{serviceId}{The obfuscated ID of the service this privilege is for}

\item{serviceName}{The name of the service this privilege is for}
}
\value{
Privilege object
}
\description{
Privilege Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for privilege resource in Directory API.
}

