% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Privileges}
\alias{Privileges}
\title{Privileges Object}
\usage{
Privileges(etag = NULL, items = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{A list of Privilege resources}
}
\value{
Privileges object
}
\description{
Privileges Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List privileges operation in Directory API.
}
