% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Role}
\alias{Role}
\title{Role Object}
\usage{
Role(Role.rolePrivileges = NULL, etag = NULL, isSuperAdminRole = NULL,
  isSystemRole = NULL, roleDescription = NULL, roleId = NULL,
  roleName = NULL, rolePrivileges = NULL)
}
\arguments{
\item{Role.rolePrivileges}{The \link{Role.rolePrivileges} object or list of objects}

\item{etag}{ETag of the resource}

\item{isSuperAdminRole}{Returns true if the role is a super admin role}

\item{isSystemRole}{Returns true if this is a pre-defined system role}

\item{roleDescription}{A short description of the role}

\item{roleId}{ID of the role}

\item{roleName}{Name of the role}

\item{rolePrivileges}{The set of privileges that are granted to this role}
}
\value{
Role object
}
\description{
Role Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for role resource in Directory API.
}
\seealso{
Other Role functions: \code{\link{Role.rolePrivileges}},
  \code{\link{directory.roles.insert}},
  \code{\link{directory.roles.patch}},
  \code{\link{directory.roles.update}}
}

