% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Role.rolePrivileges}
\alias{Role.rolePrivileges}
\title{Role.rolePrivileges Object}
\usage{
Role.rolePrivileges()
}
\value{
Role.rolePrivileges object
}
\description{
Role.rolePrivileges Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Role functions: \code{\link{Role}},
  \code{\link{directory.roles.insert}},
  \code{\link{directory.roles.patch}},
  \code{\link{directory.roles.update}}
}
