% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{RoleAssignment}
\alias{RoleAssignment}
\title{RoleAssignment Object}
\usage{
RoleAssignment(assignedTo = NULL, etag = NULL, orgUnitId = NULL,
  roleAssignmentId = NULL, roleId = NULL, scopeType = NULL)
}
\arguments{
\item{assignedTo}{The unique ID of the user this role is assigned to}

\item{etag}{ETag of the resource}

\item{orgUnitId}{If the role is restricted to an organization unit, this contains the ID for the organization unit the exercise of this role is restricted to}

\item{roleAssignmentId}{ID of this roleAssignment}

\item{roleId}{The ID of the role that is assigned}

\item{scopeType}{The scope in which this role is assigned}
}
\value{
RoleAssignment object
}
\description{
RoleAssignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for roleAssignment resource in Directory API.
}
\seealso{
Other RoleAssignment functions: \code{\link{directory.roleAssignments.insert}}
}

