% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{RoleAssignments}
\alias{RoleAssignments}
\title{RoleAssignments Object}
\usage{
RoleAssignments(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{A list of RoleAssignment resources}

\item{nextPageToken}{No description}
}
\value{
RoleAssignments object
}
\description{
RoleAssignments Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List roleAssignments operation in Directory API.
}
