% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Roles}
\alias{Roles}
\title{Roles Object}
\usage{
Roles(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{A list of Role resources}

\item{nextPageToken}{No description}
}
\value{
Roles object
}
\description{
Roles Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List roles operation in Directory API.
}
