% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Schema}
\alias{Schema}
\title{Schema Object}
\usage{
Schema(etag = NULL, fields = NULL, schemaId = NULL, schemaName = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{fields}{Fields of Schema}

\item{schemaId}{Unique identifier of Schema (Read-only)}

\item{schemaName}{Schema name}
}
\value{
Schema object
}
\description{
Schema Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Schema resource in Directory API.
}
\seealso{
Other Schema functions: \code{\link{directory.schemas.insert}},
  \code{\link{directory.schemas.patch}},
  \code{\link{directory.schemas.update}}
}

