% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{SchemaFieldSpec.numericIndexingSpec}
\alias{SchemaFieldSpec.numericIndexingSpec}
\title{SchemaFieldSpec.numericIndexingSpec Object}
\usage{
SchemaFieldSpec.numericIndexingSpec(maxValue = NULL, minValue = NULL)
}
\arguments{
\item{maxValue}{Maximum value of this field}

\item{minValue}{Minimum value of this field}
}
\value{
SchemaFieldSpec.numericIndexingSpec object
}
\description{
SchemaFieldSpec.numericIndexingSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indexing spec for a numeric field. By default, only exact match queries will be supported for numeric fields. Setting the numericIndexingSpec allows range queries to be supported.
}
\seealso{
Other SchemaFieldSpec functions: \code{\link{SchemaFieldSpec}}
}
