% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{SchemaFieldSpec}
\alias{SchemaFieldSpec}
\title{SchemaFieldSpec Object}
\usage{
SchemaFieldSpec(SchemaFieldSpec.numericIndexingSpec = NULL, etag = NULL,
  fieldId = NULL, fieldName = NULL, fieldType = NULL, indexed = NULL,
  multiValued = NULL, numericIndexingSpec = NULL, readAccessType = NULL)
}
\arguments{
\item{SchemaFieldSpec.numericIndexingSpec}{The \link{SchemaFieldSpec.numericIndexingSpec} object or list of objects}

\item{etag}{ETag of the resource}

\item{fieldId}{Unique identifier of Field (Read-only)}

\item{fieldName}{Name of the field}

\item{fieldType}{Type of the field}

\item{indexed}{Boolean specifying whether the field is indexed or not}

\item{multiValued}{Boolean specifying whether this is a multi-valued field or not}

\item{numericIndexingSpec}{Indexing spec for a numeric field}

\item{readAccessType}{Read ACLs on the field specifying who can view values of this field}
}
\value{
SchemaFieldSpec object
}
\description{
SchemaFieldSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for FieldSpec resource for Schemas in Directory API.
}
\seealso{
Other SchemaFieldSpec functions: \code{\link{SchemaFieldSpec.numericIndexingSpec}}
}
