% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Schemas}
\alias{Schemas}
\title{Schemas Object}
\usage{
Schemas(etag = NULL, schemas = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{schemas}{List of UserSchema objects}
}
\value{
Schemas object
}
\description{
Schemas Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List Schema operation in Directory API.
}

