% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Token}
\alias{Token}
\title{Token Object}
\usage{
Token(anonymous = NULL, clientId = NULL, displayText = NULL,
  etag = NULL, nativeApp = NULL, scopes = NULL, userKey = NULL)
}
\arguments{
\item{anonymous}{Whether the application is registered with Google}

\item{clientId}{The Client ID of the application the token is issued to}

\item{displayText}{The displayable name of the application the token is issued to}

\item{etag}{ETag of the resource}

\item{nativeApp}{Whether the token is issued to an installed application}

\item{scopes}{A list of authorization scopes the application is granted}

\item{userKey}{The unique ID of the user that issued the token}
}
\value{
Token object
}
\description{
Token Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for token resource in Directory API.
}

