% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Tokens}
\alias{Tokens}
\title{Tokens Object}
\usage{
Tokens(etag = NULL, items = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{A list of Token resources}
}
\value{
Tokens object
}
\description{
Tokens Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List tokens operation in Directory API.
}

