% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{User.customSchemas}
\alias{User.customSchemas}
\title{User.customSchemas Object}
\usage{
User.customSchemas()
}
\value{
User.customSchemas object
}
\description{
User.customSchemas Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Custom fields of the user.
}
\seealso{
Other User functions: \code{\link{User}},
  \code{\link{directory.users.insert}},
  \code{\link{directory.users.patch}},
  \code{\link{directory.users.update}}
}
