% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{User}
\alias{User}
\title{User Object}
\usage{
User(User.customSchemas = NULL, addresses = NULL,
  changePasswordAtNextLogin = NULL, customSchemas = NULL, emails = NULL,
  externalIds = NULL, hashFunction = NULL, id = NULL, ims = NULL,
  includeInGlobalAddressList = NULL, ipWhitelisted = NULL, name = NULL,
  notes = NULL, orgUnitPath = NULL, organizations = NULL,
  password = NULL, phones = NULL, primaryEmail = NULL, relations = NULL,
  suspended = NULL, websites = NULL)
}
\arguments{
\item{User.customSchemas}{The \link{User.customSchemas} object or list of objects}

\item{addresses}{No description}

\item{changePasswordAtNextLogin}{Boolean indicating if the user should change password in next login}

\item{customSchemas}{Custom fields of the user}

\item{emails}{No description}

\item{externalIds}{No description}

\item{hashFunction}{Hash function name for password}

\item{id}{Unique identifier of User (Read-only)}

\item{ims}{No description}

\item{includeInGlobalAddressList}{Boolean indicating if user is included in Global Address List}

\item{ipWhitelisted}{Boolean indicating if ip is whitelisted}

\item{name}{User's name}

\item{notes}{No description}

\item{orgUnitPath}{OrgUnit of User}

\item{organizations}{No description}

\item{password}{User's password}

\item{phones}{No description}

\item{primaryEmail}{username of User}

\item{relations}{No description}

\item{suspended}{Indicates if user is suspended}

\item{websites}{No description}
}
\value{
User object
}
\description{
User Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for User object in Directory API.
}
\seealso{
Other User functions: \code{\link{User.customSchemas}},
  \code{\link{directory.users.insert}},
  \code{\link{directory.users.patch}},
  \code{\link{directory.users.update}}
}
