% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserAbout}
\alias{UserAbout}
\title{UserAbout Object}
\usage{
UserAbout(contentType = NULL, value = NULL)
}
\arguments{
\item{contentType}{About entry can have a type which indicates the content type}

\item{value}{Actual value of notes}
}
\value{
UserAbout object
}
\description{
UserAbout Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for About (notes) of a user in Directory API.
}

