% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserCustomProperties}
\alias{UserCustomProperties}
\title{UserCustomProperties Object}
\usage{
UserCustomProperties()
}
\value{
UserCustomProperties object
}
\description{
UserCustomProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a set of custom properties (i.e. all fields in a particular schema)
}

