% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserEmail}
\alias{UserEmail}
\title{UserEmail Object}
\usage{
UserEmail(address = NULL, customType = NULL, primary = NULL,
  type = NULL)
}
\arguments{
\item{address}{Email id of the user}

\item{customType}{Custom Type}

\item{primary}{If this is user's primary email}

\item{type}{Each entry can have a type which indicates standard types of that entry}
}
\value{
UserEmail object
}
\description{
UserEmail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an email.
}
