% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserExternalId}
\alias{UserExternalId}
\title{UserExternalId Object}
\usage{
UserExternalId(customType = NULL, type = NULL, value = NULL)
}
\arguments{
\item{customType}{Custom type}

\item{type}{The type of the Id}

\item{value}{The value of the id}
}
\value{
UserExternalId object
}
\description{
UserExternalId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an externalId entry.
}
