% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserIm}
\alias{UserIm}
\title{UserIm Object}
\usage{
UserIm(customProtocol = NULL, customType = NULL, im = NULL,
  primary = NULL, protocol = NULL, type = NULL)
}
\arguments{
\item{customProtocol}{Custom protocol}

\item{customType}{Custom type}

\item{im}{Instant messenger id}

\item{primary}{If this is user's primary im}

\item{protocol}{Protocol used in the instant messenger}

\item{type}{Each entry can have a type which indicates standard types of that entry}
}
\value{
UserIm object
}
\description{
UserIm Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for instant messenger of an user.
}

