% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserMakeAdmin}
\alias{UserMakeAdmin}
\title{UserMakeAdmin Object}
\usage{
UserMakeAdmin(status = NULL)
}
\arguments{
\item{status}{Boolean indicating new admin status of the user}
}
\value{
UserMakeAdmin object
}
\description{
UserMakeAdmin Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON request template for setting/revoking admin status of a user in Directory API.
}
\seealso{
Other UserMakeAdmin functions: \code{\link{directory.users.makeAdmin}}
}

