% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserName}
\alias{UserName}
\title{UserName Object}
\usage{
UserName(familyName = NULL, fullName = NULL, givenName = NULL)
}
\arguments{
\item{familyName}{Last Name}

\item{fullName}{Full Name}

\item{givenName}{First Name}
}
\value{
UserName object
}
\description{
UserName Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for name of a user in Directory API.
}

