% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserOrganization}
\alias{UserOrganization}
\title{UserOrganization Object}
\usage{
UserOrganization(costCenter = NULL, customType = NULL, department = NULL,
  description = NULL, domain = NULL, location = NULL, name = NULL,
  primary = NULL, symbol = NULL, title = NULL, type = NULL)
}
\arguments{
\item{costCenter}{The cost center of the users department}

\item{customType}{Custom type}

\item{department}{Department within the organization}

\item{description}{Description of the organization}

\item{domain}{The domain to which the organization belongs to}

\item{location}{Location of the organization}

\item{name}{Name of the organization}

\item{primary}{If it user's primary organization}

\item{symbol}{Symbol of the organization}

\item{title}{Title (designation) of the user in the organization}

\item{type}{Each entry can have a type which indicates standard types of that entry}
}
\value{
UserOrganization object
}
\description{
UserOrganization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for an organization entry.
}

