% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserPhone}
\alias{UserPhone}
\title{UserPhone Object}
\usage{
UserPhone(customType = NULL, primary = NULL, type = NULL, value = NULL)
}
\arguments{
\item{customType}{Custom Type}

\item{primary}{If this is user's primary phone or not}

\item{type}{Each entry can have a type which indicates standard types of that entry}

\item{value}{Phone number}
}
\value{
UserPhone object
}
\description{
UserPhone Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a phone entry.
}

