% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserRelation}
\alias{UserRelation}
\title{UserRelation Object}
\usage{
UserRelation(customType = NULL, type = NULL, value = NULL)
}
\arguments{
\item{customType}{Custom Type}

\item{type}{The relation of the user}

\item{value}{The name of the relation}
}
\value{
UserRelation object
}
\description{
UserRelation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a relation entry.
}
