% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Users}
\alias{Users}
\title{Users Object}
\usage{
Users(etag = NULL, nextPageToken = NULL, trigger_event = NULL,
  users = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{nextPageToken}{Token used to access next page of this result}

\item{trigger_event}{Event that triggered this response (only used in case of Push Response)}

\item{users}{List of user objects}
}
\value{
Users object
}
\description{
Users Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List Users operation in Apps Directory API.
}
