% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserUndelete}
\alias{UserUndelete}
\title{UserUndelete Object}
\usage{
UserUndelete(orgUnitPath = NULL)
}
\arguments{
\item{orgUnitPath}{OrgUnit of User}
}
\value{
UserUndelete object
}
\description{
UserUndelete Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON request template to undelete a user in Directory API.
}
\seealso{
Other UserUndelete functions: \code{\link{directory.users.undelete}}
}

