% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{UserWebsite}
\alias{UserWebsite}
\title{UserWebsite Object}
\usage{
UserWebsite(customType = NULL, primary = NULL, type = NULL,
  value = NULL)
}
\arguments{
\item{customType}{Custom Type}

\item{primary}{If this is user's primary website or not}

\item{type}{Each entry can have a type which indicates standard types of that entry}

\item{value}{Website}
}
\value{
UserWebsite object
}
\description{
UserWebsite Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for a website entry.
}

