% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{VerificationCode}
\alias{VerificationCode}
\title{VerificationCode Object}
\usage{
VerificationCode(etag = NULL, userId = NULL, verificationCode = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{userId}{The obfuscated unique ID of the user}

\item{verificationCode}{A current verification code for the user}
}
\value{
VerificationCode object
}
\description{
VerificationCode Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for verification codes in Directory API.
}
