% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{VerificationCodes}
\alias{VerificationCodes}
\title{VerificationCodes Object}
\usage{
VerificationCodes(etag = NULL, items = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{A list of verification code resources}
}
\value{
VerificationCodes object
}
\description{
VerificationCodes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for List verification codes operation in Directory API.
}

